Add new command named utils. Add a flag to this command that is '-db'.
The point here is that if the command 'raytool utils -db' is executed, the tool will removed the database and will create a new database with the same name. Right now it is a first rude version that do that because I assume the database name and I hardcoded it. In the future, the name of the Liferay mysql database must be get from the portal-ext.properties file, and also the username and the password to execute the command in mysql